ExamEvluationSystem
Class contains 100 students.
Exam Contains 50 questions.
Each questions contains 4 choices (as ‘A’, ‘B’,’C’ and ‘D’) and empty answer (if not filled) (all of them have same probability)
Randomly generate an exam result.
Ramdomly generate a key.
Evaluate the exam and list the results on the screen. As follows (And also Show the maximum, minimum and average grades).